CFY-7348 Sort out plugin dependencies
The main changes to the dependencies are as follows:
1. wagon==0.6.1 - ARIA currently depends on wagon 0.6.0, which in turn depends
on the latest wheel version(which dropped support for python2.6). In order
to still support python2.6, we hard-code the wagon dependncy under the plugin
dependencies.
2. setuptools>=36.5.0 - ARIA requires a lower version of setuptools, a version
which breaks CFY. However, the CFY version of setuptools is compatible with ARIA.

Both of these dependencies should follow both the ARIA dependncy and the CFY dependency
in order to override them.